08.12 - final refactorings (step 6 - refactor the auth controller) - move the LoginResponse dto to the auth package
